I created addNewOrder and started with get all order items. Things I need to change tomorrow. Get all DB methods or state changing methods out of map(). Only simple methods can be in map. Also I have not transactional DB right now. Right now if my code fails somewhere it still commits changes done before. Autocommit is on. So tomorrow three things are definitely needed to be done: getAllOrders (I will use join for getting item name and sending back DTO) Refactor code I have already done, I think I should do that now. Then make database connection transactional. If time is left after those, just continue.

+++ restaurant/src/main/java/org/example/dto/OrderItemDTO.java
@@ -1,13 +1,15 @@
 package org.example.dto;
 
+import java.math.BigDecimal;
+
 public class OrderItemDTO {
     private final int order_id;
     private final int item_id;
     private final int quantity;
-    private final float price;
+    private final BigDecimal price;
     private final boolean ready;
 
-    public OrderItemDTO(int order_id, int item_id, int quantity, float price, boolean ready) {
+    public OrderItemDTO(int order_id, int item_id, int quantity, BigDecimal price, boolean ready) {
         this.order_id = order_id;
         this.item_id = item_id;
         this.quantity = quantity;
@@ -19,7 +21,7 @@ public class OrderItemDTO {
         return ready;
     }
 
-    public float getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
